neutron/requirements.txt
pbr>=0.6,<1.0

Paste
PasteDeploy>=1.5.0
Routes>=1.12.3
amqplib>=0.6.1
anyjson>=0.3.3
argparse
Babel>=1.3
eventlet>=0.13.0
greenlet>=0.3.2
httplib2>=0.7.5
requests>=1.1
iso8601>=0.1.9
jsonrpclib
Jinja2
kombu>=2.4.8
netaddr>=0.7.6
python-neutronclient>=2.3.4,<3
SQLAlchemy>=0.7.8,<=0.9.99
WebOb>=1.2.3
python-keystoneclient>=0.7.0
alembic>=0.4.1
six>=1.5.2
stevedore>=0.14
oslo.config>=1.2.0
oslo.rootwrap
python-novaclient>=2.17.0
taskflow
